German Style Formatting of Numbers

The label_xxx_xxx family of functions makes it easy to format numbers in decimal format, as percentages or as currencies.

Usage

label_number_de(
  accuracy = 1,
  scale = 1,
  prefix = "",
  suffix = "",
  big.mark = NULL,
  decimal.mark = NULL,
  trim = TRUE,
  ...
)

number_de(
  x,
  accuracy = 1,
  scale = 1,
  prefix = "",
  suffix = "",
  big.mark = NULL,
  decimal.mark = NULL,
  trim = TRUE,
  ...
)

label_percent_de(
  accuracy = 1,
  scale = 100,
  prefix = "",
  suffix = NULL,
  big.mark = NULL,
  decimal.mark = NULL,
  trim = TRUE,
  ...
)

percent_de(
  x,
  accuracy = 1,
  scale = 100,
  prefix = "",
  suffix = NULL,
  big.mark = NULL,
  decimal.mark = NULL,
  trim = TRUE,
  ...
)

label_currency_de(
  accuracy = 1,
  scale = 1,
  currency = "EUR",
  prefix = NULL,
  suffix = NULL,
  big.mark = NULL,
  decimal.mark = NULL,
  trim = TRUE,
  ...
)

currency_de(
  x,
  accuracy = 1,
  scale = 1,
  currency = "EUR",
  prefix = NULL,
  suffix = NULL,
  big.mark = NULL,
  decimal.mark = NULL,
  trim = TRUE,
  ...
)

Arguments

accuracy

A number to round to. Use (e.g.) 0.01 to show 2 decimal places of precision. If NULL, the default, uses a heuristic that should ensure breaks have the minimum number of digits needed to show the difference between adjacent values.

Applied to rescaled data.

scale

A scaling factor: x will be multiplied by scale before formatting. This is useful if the underlying data is very small or very large.

prefix

Additional text to display before the number. The suffix is applied to absolute value before style_positive and style_negative are processed so that prefix = "$" will yield (e.g.) -$1 and ($1).

suffix

Additional text to display after the number.

big.mark

Character used between every 3 digits to separate thousands.

decimal.mark

The character to be used to indicate the numeric decimal point.

trim

Logical, if FALSE, values are right-justified to a common width (see base::format()).

...

Other arguments passed on to base::format().

x

a numeric

currency

currency symbol

Details

label_number_de formats numbers in decimal format.

label_percent_de formats numbers as percentages.

label_currency_de formats numbers as currencies.

Examples

if (FALSE) { # \dontrun{
require(scales)
demo_continuous(
  c(-1e6, 1e6),
  labels = label_number_de()
)
demo_continuous(
  c(-1, 1),
  labels = label_percent_de(accuracy = .01)
)
demo_continuous(
  c(-1, 1),
  labels = label_currency_de(accuracy = .1)
)
} # }