Labeling Functions Factory

Creates source code to add new labeling functions

Usage

label_factory(iso2, unit, country, adjectival, params = list())

Arguments

iso2

a character. Country code to be appended to the

unit

a charatcer. One of "number", "percent, "dollar" or "euro

country

a charatcer. Country label to be used in the docs.

adjectival

a charatcer. Country adjectival to be used in the docs.

params

a list. Specification of the labeling function. Default values.

Examples

params <- list(big.mark = ".", decimal.mark = ",", suffix = "%")
label_factory("de", "percent", "Germany", "German", params = params)
#> Warning: Could not use colored = TRUE, as the package prettycode is not installed. Please install it if you want to see colored output or see `?print.vertical` for more information.
#> #' German Number Formatting
#> #'
#> #' The label_xxx_de family of functions makes it easy to label numbers
#> #'     in decimal format, as percentages or as currencies applying the number
#> #'     formatting style used in Germany which uses a "," as the decimal mark
#> #'     and a "." as the thousands seperator.
#> #'
#> #' All label_() functions return a "labelling" function, i.e. a function
#> #' that takes a vector x and returns a character vector of length(x) giving
#> #' a label for each input value.
#> #'
#> #' Use
#> #'  * `label_number_de` to format numbers in decimal format.
#> #'  * `label_percent_de` to format numbers as percentages.
#> #'  * `label_dollar_de` to format numbers as dollars.
#> #'  * `label_euro_de` to format numbers as euros.
#> #'
#> #' @inheritParams scales::label_number
#> #'
#> #' @export
#> #'
#> #' @examples
#> #' demo_continuous(c(0, 1), labels = label_percent_de())
#> label_percent_de <- function(accuracy = NULL,
#>                              scale = 1,
#>                              big_mark = ".",
#>                              decimal_mark = ",",
#>                              prefix = "",
#>                              suffix = "%",
#>                              trim = TRUE,
#>                              ...) {
#>   label_percent(accuracy = accuracy, scale = scale, big.mark = big.mark, decimal.mark = decimal.mark, trim = trim, ...)
#> }